comments: refactor
