+++ src/packages/frontend/frame-editors/code-editor/actions.ts
@@ -19,7 +19,6 @@ const WIKI_HELP_URL = "https://github.com/sagemathinc/cocalc/wiki/";
 const SAVE_ERROR = "Error saving file to disk. ";
 const SAVE_WORKAROUND =
   "Ensure your network connection is solid. If this problem persists, you might need to close and open this file, restart this project in project settings, or contact support (help@cocalc.com)";
-import { webapp_client } from "@cocalc/frontend/webapp-client";
 import type { TourProps } from "antd";
 import { delay } from "awaiting";
 import * as CodeMirror from "codemirror";
@@ -110,6 +109,7 @@ import {
   chat,
   getSideChatActions,
 } from "@cocalc/frontend/frame-editors/generic/chat";
+import { Comments } from "@cocalc/frontend/frame-editors/generic/comments";
 
 interface gutterMarkerParams {
   line: number;
@@ -205,14 +205,14 @@ export class Actions<
   private _active_id_history: string[] = [];
   private _spellcheck_is_supported: boolean = false;
 
+  public comments: Comments;
+
   // We store these actions here so that we can remove the actions
   // and store for time travel when this editor is closed.
   public timeTravelActions?: TimeTravelActions;
 
   // multifile support. this will be set to the path of the parent file (master)
   protected parent_file: string | undefined = undefined;
-
-  private commentsDB?: SyncDB = undefined;
 
   _init(
     project_id: string,
@@ -1452,9 +1452,9 @@ export class Actions<
   }
 
   // Get the underlying codemirror doc that editors are using.
-  _get_doc(): CodeMirror.Doc {
+  _get_doc = (): CodeMirror.Doc => {
     return cm_doc_cache.get_doc(this.project_id, this.path);
-  }
+  };
 
   _recent_cm(): CodeMirror.Editor | undefined {
     if (this._state === "closed") return;
@@ -3116,179 +3116,12 @@ export class Actions<
     }
   }
 
-  setComment = ({
-    pos,
-    id,
-    hide,
-    noSave,
-  }: {
-    pos?: {
-      from: { line: number; ch: number };
-      to: { line: number; ch: number };
-    };
-    id: string;
-    hide?: boolean;
-    noSave?: boolean;
-  }) => {
-    const doc = this._get_doc();
-    if (!doc) {
-      throw Error("no cm doc, so can't mark");
-    }
-    for (const mark of doc.getAllMarks()) {
-      if (mark.attributes?.style == id) {
-        if (pos == null) {
-          // also locate it
-          const pos1 = mark.find();
-          if (pos1 != null) {
-            // @ts-ignore
-            pos = pos1;
-          }
-        }
-        // overwriting existing mark, so remove that one (gets created again below)
-        mark.clear();
-      }
-    }
-    if (pos == null) {
-      throw Error("unable to find mark");
-    }
-    // create the mark
-    doc.markText(pos.from, pos.to, {
-      css: hide ? "" : "background:#fef2cd",
-      shared: true,
-      attributes: { style: id },
+  initComments = () => {
+    this.comments = new Comments({
+      getDoc: this._get_doc,
+      path: this.path,
+      project_id: this.project_id,
+      syncdoc: this._syncstring,
     });
-    if (!noSave) {
-      this.saveCommentsDebounce();
-    }
-  };
-
-  private getComments = () => {
-    const doc = this._get_doc();
-    // @ts-ignore  (TODO)
-    const time = this._syncstring.patch_list.newest_patch_time().valueOf();
-    const hash = this._syncstring.hash_of_live_version();
-    return doc
-      .getAllMarks()
-      .filter((mark) => mark.attributes?.style)
-      .map((mark) => {
-        const id = mark.attributes!.style;
-        const hide = !mark.css;
-        // @ts-ignore
-        const { from, to } = mark.find();
-        return {
-          id,
-          time,
-          hash,
-          hide,
-          pos: { from, to },
-        };
-      });
-  };
-
-  private commentsPath = () => {
-    return aux_file(this.path, "comments");
-  };
-
-  getCommentsDB = async () => {
-    if (this.commentsDB == null) {
-      this.commentsDB = await webapp_client.sync_client.sync_db({
-        project_id: this.project_id,
-        path: this.commentsPath(),
-        primary_keys: ["id"],
-        ephemeral: true,
-        cursors: true,
-      });
-      this._syncstring.on("close", () => {
-        console.log("closing comments db");
-        this.commentsDB?.close();
-      });
-      this.commentsDB.on(
-        "change",
-        debounce(
-          () => {
-            this.loadComments();
-          },
-          3000,
-          { leading: true, trailing: true },
-        ),
-      );
-      if (this.commentsDB.get_state() != "ready") {
-        await once(this.commentsDB, "ready");
-      }
-    }
-    return this.commentsDB;
-  };
-
-  private hasComments = () => {
-    try {
-      const doc = this._get_doc();
-      return (
-        doc.getAllMarks().filter((mark) => mark.attributes?.style).length > 0
-      );
-    } catch (_) {
-      // expected when no cm is open or document is closed
-      return false;
-    }
-  };
-
-  saveComments = reuseInFlight(async () => {
-    if (!this.hasComments()) {
-      return;
-    }
-    let d = 100;
-    while (this._syncstring.has_unsaved_changes()) {
-      if (d >= 30000) {
-        console.warn(
-          "something is going wrong waiting for document to stabilize",
-        );
-      }
-      await delay(d);
-      if (this._syncstring.get_state() == "closed") {
-        return;
-      }
-      d = Math.min(30000, 1.3 * d);
-      await this._syncstring.save();
-    }
-    // due to above loop, right now there are no unsaved changes, so we can safely
-    // get the comments and write them out:
-    const comments = this.getComments();
-    if (comments == null) {
-      return;
-    }
-    const db = await this.getCommentsDB();
-    for (const comment of comments) {
-      db.set(comment);
-    }
-    await db.save_to_disk();
-  });
-
-  saveCommentsDebounce = debounce(this.saveComments, 3000);
-
-  loadComments = async (force?) => {
-    const db = await this.getCommentsDB();
-    const hash = this._syncstring.hash_of_live_version();
-    for (const comment of db.get()) {
-      console.log(comment.toJS());
-      if (comment.get("hash") == hash || force) {
-        console.log("using it!");
-        const { id, hide, pos } = comment.toJS();
-        this.setComment({ pos, id, hide, noSave: true });
-      } else {
-        console.log("NOT using it -- need algorithm to transform");
-      }
-    }
-  };
-
-  initComments = async () => {
-    if (
-      await this.redux
-        .getProjectActions(this.project_id)
-        .path_exists(this.commentsPath())
-    ) {
-      // probably comments, so load them
-      await this.loadComments();
-    }
-    // also periodically save comments out when activity stops
-    this._syncstring.on("change", this.saveCommentsDebounce);
   };
 }
