+++ src/packages/frontend/frame-editors/generic/comments/index.ts
@@ -1,0 +1,1 @@
+export { Comments } from "./comments";
